ai-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
openai==1.59.6
python-multipart==0.0.20
python-dotenv==1.0.1
